adventura-backend/ml/requirements.txt
numpy
pandas
scikit-learn
flask
psycopg2-binary